chore: move vendor to utils and update license notes
