+++ LICENSE
@@ -1,6 +1,6 @@
 Faker - Copyright (c) 2022
 
-This software consists of voluntary contributions made by many individuals. 
+This software consists of voluntary contributions made by many individuals.
 For exact contribution history, see the revision history
 available at https://github.com/faker-js/faker
 
@@ -23,7 +23,7 @@ LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-=== 
+===
 
 From: https://github.com/faker-js/faker/commit/a9f98046c7d5eeaabe12fc587024c06d683800b8
 To: https://github.com/faker-js/faker/commit/29234378807c4141588861f69421bf20b5ac635e
